Fix errors Student 2

+++ src/main/java/acme/forms/ClientDashboard.java
@@ -18,16 +18,16 @@ public class ClientDashboard extends AbstractForm {
 	// Attributes -------------------------------------------------------------
 
 	// total number of progress logs with a completeness rate below 25%
-	double						totalNumProgressLogLessThan25;
+	int							totalNumProgressLogLessThan25;
 
 	// total number of progress logs with a completeness rate between 25% and 50%
-	double						totalNumProgressLogBetween25And50;
+	int							totalNumProgressLogBetween25And50;
 
 	// total number of progress logs with a completeness rate between 50% and 75%
-	double						totalNumProgressLogBetween50And75;
+	int							totalNumProgressLogBetween50And75;
 
 	// total number of progress logs with a completeness rate above 75%
-	double						totalNumProgressLogAbove75;
+	int							totalNumProgressLogAbove75;
 
 	// average, deviation, minimum, and maximum budget of the contracts
